+++ src/main/java/com/abcdedu_backend/memberv2/application/MemberInfoService.java
@@ -69,11 +69,6 @@ public class MemberInfoService implements MemberInfoUseCase{
     }
 
     @Override
-    public void checkDuplicateEmail(String toEmail) {
-
-    }
-
-    @Override
     @Transactional
     public void updatePassword(Long memberId, String newPassword) {
         String newEncodedPassword = passwordEncoder.encode(newPassword);
