+++ src/main/java/com/abcdedu_backend/memberv2/application/domain/EmailCode.java
@@ -1,0 +1,7 @@
+package com.abcdedu_backend.memberv2.application.domain;
+
+public record EmailCode (
+        String email,
+        String code
+){
+}
